Create roulette1.py
Rouletter comes up with random strategies to try and beat the game of Roulette.

It assumes the player will always bet on even, and will double the bet each time they lose.

It then iterates around this strategy to try and come up with a way to beat the game.